fix: add minimax_tokens to cleanup and fix company_id type
- Add minimax_tokens to tables to drop
- Change company_id from unsignedBigInteger to unsignedInteger to match companies table

+++ database/migrations/2025_07_26_110000_create_minimax_tokens_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('minimax_tokens', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('company_id');
+            $table->unsignedInteger('company_id'); // companies.id is unsigned int, not bigint
             $table->string('token_name');
             $table->text('api_token'); // Encrypted token storage
             $table->boolean('is_active')->default(true);
